refactor: move ZmqSink to nes federate
