+++ fed/nes-federate/src/main/java/com/github/paguos/mosaic/fed/nebulastream/stream/ZmqSink.java
@@ -1,16 +1,10 @@
-package com.github.paguos.mosaic.app;
+package com.github.paguos.mosaic.fed.nebulastream.stream;
 
-import com.github.paguos.mosaic.fed.ambassador.NesController;
-import com.github.paguos.mosaic.fed.msg.SerializableDataType;
 import com.github.paguos.mosaic.fed.msg.SerializableSchema;
-import com.github.paguos.mosaic.fed.nebulastream.stream.Schema;
-import com.github.paguos.mosaic.fed.nebulastream.stream.SchemaParser;
 import com.google.protobuf.InvalidProtocolBufferException;
-import org.eclipse.mosaic.rti.api.InternalFederateException;
 import org.zeromq.SocketType;
 import org.zeromq.ZContext;
 import org.zeromq.ZMQ;
-import org.zeromq.ZMsg;
 
 import java.util.Arrays;
 import java.util.concurrent.ArrayBlockingQueue;
